+++ tests/test_edge_cases.py
@@ -1,6 +1,6 @@
 # Copyright 2024, Battelle Energy Alliance, LLC All Rights Reserved.
 import copy
-import math
+import io
 import montepy
 from montepy.errors import *
 import os
@@ -41,13 +41,9 @@ class EdgeCaseTests(TestCase):
             montepy.input_parser.block_type.BlockType.DATA,
         )
         problem.data_inputs.append(montepy.data_inputs.data_parser.parse_data(card))
-        try:
-            with self.assertRaises(MalformedInputError):
-                problem.write_to_file("out")
-        finally:
-            if os.path.exists("out"):
-                pass
-                os.remove("out")
+        with io.StringIO() as stream:
+            with pytest.raises(MalformedInputError):
+                problem.write_problem(stream)
 
     def test_shortcuts_in_special_comment(self):
         in_str = "fc247 experiment in I24 Cell Specific Heating"
@@ -187,3 +183,33 @@ def test_complex_cell_parsing(lines):
         montepy.input_parser.block_type.BlockType.CELL,
     )
     montepy.Cell(input)
+
+
+@pytest.mark.filterwarnings("ignore")
+def test_universe_after_comment():
+    problem = montepy.read_input(
+        os.path.join("tests", "inputs", "test_universe_data.imcnp")
+    )
+    problem.print_in_data_block["u"] = False
+    problem.print_in_data_block["vol"] = False
+    universes = [cell.universe.number for cell in problem.cells]
+    try:
+        out_file = "universe_after_comment"
+        problem.write_to_file(out_file)
+        # TODO check that the Universe does show up
+        found = False
+        with open(out_file, "r") as fh:
+            for line in fh:
+                print(line.rstrip())
+                assert "c IMP:n=0.0" not in line
+                if "U=" in line:
+                    found = True
+        assert found
+        problem = montepy.read_input(out_file)
+        new_universes = [cell.universe.number for cell in problem.cells]
+        assert universes == new_universes
+    finally:
+        try:
+            os.remove(out_file)
+        except FileNotFoundError:
+            pass