+++ doc/source/changelog.rst
@@ -3,6 +3,9 @@ MontePy Changelog
 
 #Next Version#
 --------------
+**Features Added**
+
+* Write problems to either file paths or streams (file handles) with MCNP_Problem.write_problem() (:issue:`492`).
 
 **Features Added**
 
@@ -11,6 +14,7 @@ MontePy Changelog
 **Bug Fixes**
 
 * Fixed bug that didn't show metastable states for pretty printing and isotope. Also handled the case that Am-241 metstable states break convention (:issue:`486`).
+
 
 0.3.3
 --------------