+++ doc/source/changelog.rst
@@ -6,6 +6,7 @@ MontePy Changelog
 **Features Added**
 
 * Write problems to either file paths or streams (file handles) with MCNP_Problem.write_problem() (:issue:`492`).
+* When adding a material to problem.materials it will also be added to problem.data_inputs, ensuring it is printed to the file (:pull:`488`).
 
 **Bug Fixes**