+++ doc/source/changelog.rst
@@ -3,10 +3,16 @@ MontePy Changelog
 
 #Next Version#
 --------------
+**Features Added**
+
+* Write problems to either file paths or streams (file handles) with MCNP_Problem.write_problem() (:issue:`492`).
+* When adding a material to problem.materials it will also be added to problem.data_inputs, ensuring it is printed to the file (:pull:`488`).
 
 **Bug Fixes**
 
 * Fixed bug that didn't show metastable states for pretty printing and isotope. Also handled the case that Am-241 metstable states break convention (:issue:`486`).
+* Fixed bug where cell modifiers could be made irrelevant by being added after a comment (:issue:`483`).
+
 
 0.3.3
 --------------